requirements.txt
alembic==1.11.1
anyio==3.6.2
asyncpg==0.27.0
bcrypt==4.0.1
certifi==2023.5.7
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.2
httptools==0.5.0
httpx==0.24.1
idna==3.4
Mako==1.2.4
MarkupSafe==2.1.2
numpy==1.24.3
pandas==2.0.2
passlib==1.7.4
psycopg2==2.9.6
pyasn1==0.5.0
pydantic==1.10.7
pypng==0.20220715.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-telegram-bot==20.3
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
SQLAlchemy-Utils==0.41.1
starlette==0.27.0
typing_extensions==4.6.0
tzdata==2023.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
XlsxWriter==3.1.2
